requirements.txt
azure-common==1.1.28
azure-core==1.36.0
azure-identity==1.25.1
azure-mgmt-core==1.6.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==24.0.0
azure-storage-blob==12.27.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
numpy==2.3.4
oauthlib==3.3.1
pandas==2.3.3
psycopg2==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
